Add handling for al Tier 2 upgrade states either from Tier 0 or Tier 1
