+++ app/src/Modals/MyAccount/myAccount.scss
@@ -12,6 +12,7 @@
     &-attention {
       font-size: 1.8rem;
       color: $black;
+      padding-top: 1rem;
     }
   }
